### 2023-08-09 #5746 @stevenshave
Hello! I will be making an update to the publication to include benchmarking results, address the missing DOIs and fix API docs imminently.  Many thanks, Steve
