### 2023-08-10 5e0f5eb
Expanded paper, README.md and API docs
