+++ docsource/modules.rst
@@ -1,7 +1,4 @@
-fepops
-======
-
 .. toctree::
    :maxdepth: 4
-   
+
    fepops
