+++ docsource/fepops.rst
@@ -1,45 +1,0 @@
-fepops package
-==============
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   fepops.fepops_persistent
-
-Submodules
-----------
-
-fepops.benchmark module
------------------------
-
-.. automodule:: fepops.benchmark
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-fepops.dude\_preprocessor module
---------------------------------
-
-.. automodule:: fepops.dude_preprocessor
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-fepops.fepops module
---------------------
-
-.. automodule:: fepops.fepops
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: fepops
-   :members:
-   :undoc-members:
-   :show-inheritance:
