+++ .github/workflows/documentation.yaml
@@ -10,7 +10,7 @@ jobs:
       - uses: actions/setup-python@v3
       - name: Install dependencies
         run: |
-          pip install sphinx sphinx_rtd_theme recommonmark
+          pip install sphinx sphinx_rtd_theme recommonmark myst_parser
       - name: Sphinx build
         run: |
           sphinx-build docsource docs
