### 2023-08-14 116e284
Reworked benchmarking into Jupyter notebook
