### 2023-08-14 7acb83b
Updated paper with dude reference and small fixes
