+++ paper.bib
@@ -122,3 +122,14 @@
   year={2009},
   publisher={Elsevier}
 }
+
+@article{mysinger2012directory,
+  title={Directory of useful decoys, enhanced (DUD-E): better ligands and decoys for better benchmarking},
+  author={Mysinger, Michael M and Carchia, Michael and Irwin, John J and Shoichet, Brian K},
+  journal={Journal of medicinal chemistry},
+  volume={55},
+  number={14},
+  pages={6582--6594},
+  year={2012},
+  publisher={ACS Publications}
+}
