### 2023-08-14 b3f04ed
Corrected paper wording around failed structures
