+++ environment.yml
@@ -3,127 +3,47 @@ channels:
   - defaults
 dependencies:
   - _libgcc_mutex=0.1=main
-  - beautifulsoup4=4.11.1=py310h06a4308_0
-  - brotlipy=0.7.0=py310h7f8727e_1002
-  - bzip2=1.0.8=h7b6447c_0
-  - ca-certificates=2023.01.10=h06a4308_0
-  - certifi=2022.12.7=py310h06a4308_0
-  - cffi=1.15.0=py310hd667e15_1
-  - chardet=4.0.0=py310h06a4308_1003
-  - charset-normalizer=2.0.4=pyhd3eb1b0_0
-  - conda=23.1.0=py310h06a4308_0
-  - conda-build=3.23.3=py310h06a4308_0
-  - conda-package-handling=2.0.2=py310h06a4308_0
-  - conda-package-streaming=0.7.0=py310h06a4308_0
-  - cryptography=39.0.1=py310h9ce1e76_0
-  - filelock=3.9.0=py310h06a4308_0
-  - glob2=0.7=pyhd3eb1b0_0
-  - icu=58.2=he6710b0_3
-  - idna=3.4=py310h06a4308_0
-  - jinja2=3.1.2=py310h06a4308_0
+  - ca-certificates=2023.05.30=h06a4308_0
   - ld_impl_linux-64=2.38=h1181459_1
-  - libarchive=3.5.2=h5de8990_0
   - libffi=3.3=he6710b0_2
   - libgcc-ng=9.1.0=hdf63c60_0
-  - liblief=0.11.5=h295c915_1
   - libstdcxx-ng=9.1.0=hdf63c60_0
-  - libuuid=1.0.3=h7f8727e_2
-  - libxml2=2.9.14=h74e7548_0
-  - lz4-c=1.9.3=h295c915_1
-  - markupsafe=2.1.1=py310h7f8727e_0
   - ncurses=6.3=h7f8727e_2
-  - openssl=1.1.1t=h7f8727e_0
-  - patch=2.7.6=h7b6447c_1001
-  - patchelf=0.13=h295c915_0
-  - pip=23.0.1=py310h06a4308_0
-  - pkginfo=1.9.6=py310h06a4308_0
-  - pluggy=1.0.0=py310h06a4308_1
-  - psutil=5.8.0=py310h7f8727e_1
-  - py-lief=0.11.5=py310h295c915_1
-  - pycosat=0.6.3=py310h7f8727e_0
-  - pycparser=2.21=pyhd3eb1b0_0
-  - pyopenssl=23.0.0=py310h06a4308_0
-  - pysocks=1.7.1=py310h06a4308_0
-  - python=3.10.4=h12debd9_0
-  - python-libarchive-c=2.9=pyhd3eb1b0_1
-  - pytz=2022.7=py310h06a4308_0
-  - pyyaml=6.0=py310h7f8727e_0
+  - openssl=1.1.1v=h7f8727e_0
+  - pip=23.2.1=py39h06a4308_0
+  - python=3.9.12=h12debd9_1
   - readline=8.1.2=h7f8727e_1
-  - requests=2.28.1=py310h06a4308_1
-  - ripgrep=13.0.0=hbdeaff8_0
-  - ruamel.yaml=0.16.10=py310h7f8727e_1
-  - ruamel.yaml.clib=0.2.6=py310h7f8727e_0
-  - setuptools=65.6.3=py310h06a4308_0
-  - six=1.16.0=pyhd3eb1b0_1
-  - soupsieve=2.3.2.post1=py310h06a4308_0
+  - setuptools=68.0.0=py39h06a4308_0
[…]
