### 2023-08-15 a989a15
Minor edits to the paper and readme
