+++ paper.bib
@@ -7,6 +7,7 @@
   pages={622--633},
   year={2007},
   publisher={Elsevier}
+  doi={10.1016/j.jmgm.2007.02.005}
 }
 
 @article{jenkins20043d,
@@ -18,6 +19,7 @@
   pages={6144--6159},
   year={2004},
   publisher={ACS Publications}
+  doi = {10.1021/jm049654z}
 }
 
 @article{jenkins2013feature,
@@ -27,6 +29,7 @@
   pages={155--174},
   year={2013},
   publisher={Wiley Online Library}
+  doi = {10.1002/9783527665143.ch10}
 }
 
 @article{ripphausen2011state,
@@ -38,6 +41,8 @@
   pages={372--376},
   year={2011},
   publisher={Elsevier}
+  doi = {10.1016/j.drudis.2011.02.011}
+
 }
 
 @article{hughes2011principles,
@@ -49,6 +54,7 @@
   pages={1239--1249},
   year={2011},
   publisher={Wiley Online Library}
+  doi = {10.1111/j.1476-5381.2010.01127.x}
 }
 
 @article{cortes2020qsar,
@@ -60,6 +66,7 @@
   pages={41},
   year={2020},
   publisher={Springer}
+  doi = {10.1186/s13321-020-00444-5}
 }
 
 @article{baber2006use,
@@ -71,6 +78,7 @@
   pages={277--288},
   year={2006},
   publisher={ACS Publications}
+  doi = {10.1021/ci050296y}
 }
 
 @article{mysinger2012directory,
@@ -82,6 +90,7 @@
   pages={6582--6594},
   year={2012},
   publisher={ACS Publications}
+  doi = {10.1021/jm300687e}
 }
 
 @article{landrum2013rdkit,
@@ -91,6 +100,8 @@
   volume={8},
   pages={31},
   year={2013}
+  doi = {}
+
 }
 
 @article{schreyer2012usrcat,
@@ -102,6 +113,7 @@
   pages={1--12},
   year={2012},
   publisher={BioMed Central}
+  doi = {10.1186/1758-2946-4-27}
 }
 
 @inproceedings{arthur2007k,
@@ -121,15 +133,5 @@
   pages={3336--3341},
   year={2009},
   publisher={Elsevier}
-}
-
-@article{mysinger2012directory,
-  title={Directory of useful decoys, enhanced (DUD-E): better ligands and decoys for better benchmarking},
-  author={Mysinger, Michael M and Carchia, Michael and Irwin, John J and Shoichet, Brian K},
-  journal={Journal of medicinal chemistry},
-  volume={55},
-  number={14},
-  pages={6582--6594},
-  year={2012},
-  publisher={ACS Publications}
+  doi = {10.1016/j.eswa.2008.01.039}
 }
