### 2023-08-15 #5746 @stevenshave
Hi @richardjgowers, I have now made updates to the repo including API docs, userguide and added benchmark results to the paper.  Suggested DOIs have been added to my paper.bib, but some referenced articles have no valid DOI.  Please let me know the next steps.  Many thanks, Steve
