### 2023-08-15 e1c5b85
Added missing commas before doi
