+++ paper.bib
@@ -6,7 +6,7 @@
   number={3},
   pages={622--633},
   year={2007},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi={10.1016/j.jmgm.2007.02.005}
 }
 
@@ -18,7 +18,7 @@
   number={25},
   pages={6144--6159},
   year={2004},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
   doi = {10.1021/jm049654z}
 }
 
@@ -28,7 +28,7 @@
   journal={Scaffold hopping in medicinal chemistry},
   pages={155--174},
   year={2013},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
   doi = {10.1002/9783527665143.ch10}
 }
 
@@ -40,9 +40,8 @@
   number={9-10},
   pages={372--376},
   year={2011},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi = {10.1016/j.drudis.2011.02.011}
-
 }
 
 @article{hughes2011principles,
@@ -53,7 +52,7 @@
   number={6},
   pages={1239--1249},
   year={2011},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
   doi = {10.1111/j.1476-5381.2010.01127.x}
 }
 
@@ -65,7 +64,7 @@
   number={1},
   pages={41},
   year={2020},
-  publisher={Springer}
+  publisher={Springer},
   doi = {10.1186/s13321-020-00444-5}
 }
 
@@ -77,7 +76,7 @@
   number={1},
   pages={277--288},
   year={2006},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
   doi = {10.1021/ci050296y}
 }
 
@@ -89,7 +88,7 @@
   number={14},
   pages={6582--6594},
   year={2012},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
   doi = {10.1021/jm300687e}
 }
 
@@ -100,8 +99,6 @@
   volume={8},
   pages={31},
   year={2013}
-  doi = {}
-
 }
 
 @article{schreyer2012usrcat,
@@ -112,7 +109,7 @@
   number={1},
   pages={1--12},
   year={2012},
-  publisher={BioMed Central}
+  publisher={BioMed Central},
   doi = {10.1186/1758-2946-4-27}
 }
 
@@ -132,6 +129,6 @@
   number={2},
   pages={3336--3341},
   year={2009},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi = {10.1016/j.eswa.2008.01.039}
 }
