+++ environment.yml
@@ -23,7 +23,7 @@ dependencies:
   - pip:
     - build==0.10.0
     - fast-pytorch-kmeans==0.1.9
-    - fepops==1.4.0
+    - fepops
     - fire==0.5.0
     - joblib==1.3.2
     - numpy==1.22.4
