### 2023-08-16 #5763 @editorialbot
[REVIEW]: OpenFEPOPS: A Python implementation of the FEPOPS molecular similarity technique
**Submitting author:** @stevenshave (<a href="http://orcid.org/0000-0001-6996-3663">Steven Shave</a>)
**Repository:** https://github.com/JustinYKC/FEPOPS
**Branch with paper.md** (empty if default branch):
**Version:** 1.4.0
