### 2023-08-20 71658d0
Fixed bug that always assigned None to mol variable
