### 2023-08-31 JustinYKC/FEPOPS#11 @exs-cbouy
[Functionality] Lack of feature scaling for Kmedoids
Refers to the JOSS submission review: https://github.com/openjournals/joss-reviews/issues/5763

When performing distance-based clustering using different types of features, it is highly recommended to scale the features beforehand, otherwise some features may be given more importance than others on the sole basis that their absolute values are larger numbers, instead of treating all features with the same importance.

This is not a problem when using KMeans for the clustering of atoms because all the "features" are atomic coordinates so the units and scale are the same.

For the KMedoids step however, the features are FEPOPS, which includes different objects: distances, sum of partial charges, sum of logP, binary flags for the presence of HBond donors and acceptors. These all operate on different scales, so by omitting a standardisation step before the clustering with KMedoids, it is very likely that the cluster centroids will mostly be selected based on the distance-based features and not so much based on the HBond-based features, e.g. given 2 conformers, the difference in values for an HBond-feature will be at most 1, whereas for a distance-feature it is usually going to be much larger than 1, so the clustering will be biased towards distance-based features.

Currently the feature scaling is done **after** KMedoids, during the similarity calculation, which I consider to be a critical omission. The original paper describes this feature scaling procedure in the Kmedoids step as well (see the *Selection of Representative FEPOPS Conformer* section in the *Theory and Methods* section of the 2004 paper) so I strongly recommend the authors to adjust their implementation accordingly, i.e. move the feature scaling to the FEPOPS calculation rather than the similarity one.

I would also recommend to rerun the benchmark once that modification is made, as it may impact which conformers are selected (likely improving the results for FEPOPS in the benchmark).

This is the only thing blocking me from ticking the `Functionality` checkbox in the review as it differs from the original implementation without justification, otherwise the rest of the implementation seem valid.
