### 2023-08-31 JustinYKC/FEPOPS#12 @exs-cbouy
[Documentation] Statement of need missing from docs
Refers to the JOSS submission review: https://github.com/openjournals/joss-reviews/issues/5763

The statement of need that is visible in the paper is missing from the docs and is required to be present in the JOSS checklist.
