### 2023-08-31 JustinYKC/FEPOPS#13 @exs-cbouy
[Documentation] Lack of community guidelines
Refers to the JOSS submission review: https://github.com/openjournals/joss-reviews/issues/5763

The documentation is missing the following item required by JOSS:
> Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

This usually comes in the form of a `CONTRIBUTING.md` file in the repo that you can then include in the docs. An example of which can be found on the [GitHub docs](https://github.com/github/docs/blob/main/CONTRIBUTING.md).
