### 2023-08-31 JustinYKC/FEPOPS#14 @exs-cbouy
[Documentation] Tests
Refers to the JOSS submission review: https://github.com/openjournals/joss-reviews/issues/5763

The source repository contains tests but they are not automatically run by a CI solution, one has to manually run them to confirm that the installation is working correctly and/or that any modification made to the implementation did not break anything.
In its current form the documentation is lacking some instructions on how to run those tests (i.e. pip install pytest and run `pytest tests/`)

I'll also include a minor comment on the tests themselves (optional for you to tackle): in the last test, [that line](https://github.com/JustinYKC/FEPOPS/blob/master/test/tests.py#L62) is not really checking that the results for the persistent fepops storage object are the same, it only compares the keys (molecule names) instead of also comparing the values. It should just be `assert fepops_dict == fepops_dict2`
