### 2023-08-31 JustinYKC/FEPOPS#15 @exs-cbouy
[Documentation] API docs section not compiled correctly
Refers to the JOSS submission review: https://github.com/openjournals/joss-reviews/issues/5763

The [API section](https://justinykc.github.io/FEPOPS/fepops.html) seem to not be compiled correctly as it is not showing the content of the docstrings for the different classes/functions which is a shame since these are very well documented.

Without this it's hard for me to justify that the core functionality of the software is documented to a satisfactory level as users would have trouble accessing it unless they read the source code.
