### 2023-08-31 JustinYKC/FEPOPS#16 @exs-cbouy
[Software paper] Lacking mention of how features are computed
Refers to the JOSS submission review: https://github.com/openjournals/joss-reviews/issues/5763

The paper is missing one or two sentences on how the features are computed:

- partial charges: Gasteiger charges calculated with RDKit
- Atomic logP contribution: Crippen method as implemented in RDKit
- Hbond donors and acceptors: SMARTS patterns taken from DAYLIGHT

Since the SMARTS pattern are taken from the [DAYLIGHT website](https://www.daylight.com/dayhtml_tutorials/languages/smarts/smarts_examples.html#H_BOND), they must be added to the references. (and ideally the [RDKit](https://www.rdkit.org/docs/Overview.html#citing-the-rdkit) as well)
