### 2023-08-31 275da17
Added contributing and code of conduct markdown
