### 2023-08-31 JustinYKC/FEPOPS#12 @stevenshave
Thank you for reviewing our package. I am working through the raised issues and will build a new pypi package once completed.
For this issue I have added the statement of need to the README.md, which is now imported into the docs when built upon commit. This addition was made in 0bffab8.
