### 2023-08-31 #5763 @exs-cbouy
@richardjgowers as a reviewer am I supposed to rerun the benchmarking notebook (as it would qualify as "original results") and make sure that the results are the same, or is making sure that the notebook is runnable enough? It's quite slow to run (but the notebook works).
