### 2023-08-31 JustinYKC/FEPOPS#13 @stevenshave
Thank you for reviewing our package. I am working through the raised issues and will build a new pypi package once completed.

To address this issue I have added a CONTRIBUTING.md and a CODE_OF_CONDUCT.md. These were added in commit 275da17.
