### 2023-08-31 ed22de2
Fixed dependencies for install and dev install
