+++ src/fepops/fepops_persistent/fepops_persistent_abc.py
@@ -439,7 +439,9 @@ class FepopsPersistentAbstractBaseClass(metaclass=ABCMeta):
                 return np.nan
         if any(x is None for x in (fepops_features_1, fepops_features_2)):
             return np.nan
-        score = self.openfepops_object.calc_similarity(fepops_features_1, fepops_features_2)
+        score = self.openfepops_object.calc_similarity(
+            fepops_features_1, fepops_features_2
+        )
         return score if score is not None else np.nan
 
     def write(self):
