### 2023-08-31 cd29182
Changed test location and addded to pyproject.toml
