### 2023-09-01 6280684
Expanded atomic property calculation descriptions
Addresses issue #16
