### 2023-09-03 199ee51
Updated pre-score scaling means and stds
After fixing scaling for kmedoid selection, the DUDE diversity set was
reprofiled to collecting pre-score scaling means and stds.
Benchmarking then rerun.
