+++ src/fepops/fepops.py
@@ -102,53 +102,54 @@ class OpenFEPOPS:
         num_fepops_per_mol: int = 7,
         num_centroids_per_fepop: int = 4,
         descriptor_means: Tuple[float, ...] = (
-            -0.28932319,
-            0.5166312,
-            0.37458883,
-            0.99913668,
-            -0.04193182,
-            1.03616917,
-            0.27327129,
-            0.99839024,
-            0.09701198,
-            1.12969387,
-            0.23718642,
-            0.99865705,
-            0.35968991,
-            0.6649304,
-            0.4123743,
-            0.99893657,
-            5.70852885,
-            6.3707943,
-            6.47354071,
-            6.26385429,
-            6.19229367,
-            6.22946713,
+            -0.28971602,
+            0.5181022,
+            0.37487135,
+            0.99922747,
+            -0.04187301,
+            1.03382471,
+            0.27407036,
+            0.99853436,
+            0.09725517,
+            1.12824307,
+            0.23735556,
+            0.99882914,
+            0.35977538,
+            0.66653514,
+            0.41238282,
+            0.99902545,
+            5.71261449,
+            6.37716992,
+            6.47293777,
+            6.26134733,
+            6.20354385,
+            6.23201498
         ),
+        
         descriptor_stds: Tuple[float, ...] = (
-            0.35067291,
-            1.00802116,
-            0.48380817,
-            0.02926675,
-            0.15400475,
-            0.86220776,
-            0.44542581,
-            0.03999429,
-            0.16085455,
-            0.92042695,
-            0.42515847,
-            0.03655217,
-            0.35778578,
-            1.36108994,
-            0.49210665,
-            0.03252466,
-            1.96446927,
-            2.30792259,
-            2.5024708,
-            2.4155645,
-            2.29434487,
-            2.31437527,
-        ),
+            0.35110473,
+            1.00839329,
+            0.4838859,
+            0.02769204,
+            0.15418035,
+            0.86446056, 
+            0.44583626,
+            0.0381767,
+            0.16095862,
+            0.92079483,
+            0.42526185,
+            0.03413741,
+            0.35756229,
+            1.36093993,
+            0.4921059,
+            0.0311619,
+            1.9668792,
+            2.31266486,
+            2.50699385,
+            2.41269982,
+            2.30018205,
+            2.31527129
+            )
     ):
         """OpenFEPOPS (Feature Points) molecular similarity object
 
