+++ README.md
@@ -15,13 +15,13 @@ OpenFEPOPS has been benchmarked alongside some commonly used fingerprints agains
 
 | Target | Morgan 2 | MACCS | RDKit |OpenFEPOPS|
 |--------:|----------:|-------:|-------:|------------:|
-|akt1  |0.836|0.741 |0.833 |0.829|
+|akt1  |0.836|0.741 |0.833 |0.831|
 |ampc  |0.784|0.673|0.660 |0.639|
-|cp3a4 |0.603|0.582|0.613 |0.650|
+|cp3a4 |0.603|0.582|0.613 |0.647|
 |cxcr4 |0.697|0.854|0.592 |0.899|
 |gcr   |0.670|0.666|0.708  |0.616|
 |hivpr |0.780|0.681|0.759 |0.678|
-|hivrt |0.651|0.670 |0.660 |0.584|
+|hivrt |0.651|0.670 |0.660 |0.582|
 |kif11 |0.763|0.668 |0.672  |0.713|
 
 *Macro averaged AUROC scores by target and molecular similarity technique for the DUDE diversity set. Across all datasets, 19 small molecules out of 112,796 were excluded from analysis mainly due to issues in parsing to valid structures using RDKit.  See 'Explore_DUDE_diversity_set.ipynb' for further information*
