### 2023-09-03 f0a9f09
Updated tests and added to CI workflow
- Test fixed to not only compare dictionary keys
- Added CI workflow to run pytest tests and black formatter
- Added instructions to run tests to README.md
