+++ .github/workflows/ci.yaml
@@ -1,0 +1,40 @@
+# Runs the black formatter and pytest
+name: OpenFEPOPS CI
+
+on:
+  push:
+    branches: [main, development]
+  pull_request:
+    branches: [main, development]
+
+jobs:
+  build:
+    strategy:
+      matrix:
+        python-version: [3.11]
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v2
+        with:
+          fetch-depth: 0
+          
+      - name: Checkout the appropriate branch
+        run: git checkout ${{ env.BRANCH }}
+              
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v1
+        with:
+          python-version: ${{ matrix.python-version }}
+          
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -e .[dev]
+          
+      - name: Run black formatter
+        run: |
+          black .
+      - name: Run tests
+        run: |
+          pytest
