### 2023-09-03 01ff975
Corrected mol_filter.py return types
- Used typing.Optional, partially addressing issues raised in #18
