### 2023-09-03 acd95fb
Fixed docstring error
- OpenFEPOPS descriptor_stds docstring entry corrected.
- Contributes to addressing #18
