### 2023-09-03 fb7214b
Fixed API doc rendering
- Addressed #15
