+++ src/fepops/fepops_persistent/__init__.py
@@ -1,3 +1,4 @@
+"""fepops_persistent module contains functionality to cache/save FEPOPS descriptors to a file"""
 from .fepopsdb_json import FepopsDBJSON
 from .fepopsdb_sqlite import FepopsDBSqlite
 from .utils import get_persistent_fepops_storage_object
