+++ src/fepops/__init__.py
@@ -1,1 +1,3 @@
+"""FEPOPS module containing the OpenFEPOPS class"""
+
 from .fepops import OpenFEPOPS
