+++ docsource/modules.rst
@@ -3,3 +3,4 @@
    :maxdepth: 4
 
    fepops
+   fepops.fepops_persistent
