### 2023-09-03 c5e635a
Reorganised docs
- Partially addressing #18
