+++ docsource/index.rst
@@ -17,20 +17,19 @@ valuable small molecules. This implementation was recreated following the origin
 
 Source code is available on the GitHub repo here: `OpenFEPOPS GitHub Repository <https://github.com/JustinYKC/FEPOPS>`_
 
-An online version of the GitHub README.md can be found here which documents how to use the package for
-calculation of descriptors and subsequent molecular similarity calculations and is available here: `readme`
-
-.. toctree::
-   :maxdepth: 4
-   :caption: API documentation:
-
-   modules
+An online version of the GitHub README.md can be found broken down into sections below, detailing how to use the package for calculation of descriptors and subsequent molecular similarity calculations.
 
 .. toctree::
    :maxdepth: 4
    :caption: Further information:
 
    readme
+
+.. toctree::
+   :maxdepth: 4
+   :caption: API documentation:
+
+   modules
 
 Indices and tables
 ==================
