### 2023-09-04 b589376
Fixed return types in mol_file.py
