### 2023-09-04 66dbbba
Moved mol_filter.py to a seperated directory
