+++ benchmark/mol_filter.py
@@ -1,0 +1,227 @@
+from rdkit import Chem
+from rdkit.Chem import rdMolDescriptors
+from rdkit.Chem.MolStandardize import rdMolStandardize
+import pandas as pd
+from typing import Optional
+
+
+class Filter:
+    """Filter object allows filtering of molecules by property
+
+    Filter object allowing filtering by min atoms, max rings, and max rotatable bonds in a molecule.
+    """
+
+    def __init__(
+        self, min_atoms: int = 4, max_rings: int = 9, max_n_rot: int = 40
+    ) -> None:
+        self.min_atoms = min_atoms
+        self.max_rings = max_rings
+        self.max_n_rot = max_n_rot
+
+    def __call__(self, mol: Chem.rdchem.Mol) -> Optional[Chem.rdchem.Mol]:
+        """Apply all filters
+
+        Parameters
+        ----------
+        mol : Chem.rdchem.Mol
+          The Rdkit mol object of the input molecule.
+
+        Returns
+        -------
+        filter_result : Optional[Chem.rdchem.Mol]
+          Return a Rdkit mol object if the input molecule meets all criteria, otherwise return None.
+        """
+        filter_result = self.filter_mol(mol)
+        return filter_result
+
+    def filter_mol(self, mol: Chem.rdchem.Mol) -> Optional[Chem.rdchem.Mol]:
+        """Apply all filters
+
+        Perform four filters (minimum number of atoms, maximum number of rings, maximum number of rotatable bonds,
+        and salt/ion) in sequence to obtain the final molecules as required.
+
+        Parameters
+        ----------
+        mol : Chem.rdchem.Mol
+          The Rdkit mol object of the input molecule.
+
+        Returns
+        -------
+        Optional[Chem.rdchem.Mol]
+          Return a Rdkit mol object if the input molecule meets all criteria, otherwise return None.
+        """
+        mol = self.atom_num(mol, self.min_atoms)
+        if mol is None:
+            return None
+        else:
+            mol = self.ring_num(mol, self.max_rings)
+            if mol is None:
+                return None
+            else:
+                mol = self.rotation_bond_num(mol, self.max_n_rot)
+                if mol is None:
+                    return None
+                else:
+                    mol = self.remove_salt(mol)
+                    if mol is None:
+                        return None
+                    else:
+                        return mol
+
+    def read_as_mol(self, query_smiles: str) -> Chem.rdchem.Mol:
+        """Read SMILES strings from a file
+
+        Read molecular SMILES strings as Rdkit mol objects from a file.
+
+        Parameters
+        ----------
+        query_smiles : str
+          A molecular SMILES string.
+
+        Returns
+        -------
+        Chem.rdchem.Mol
+          Return a Rdkit mol object.
+        """
+        return Chem.MolFromSmiles(query_smiles)
+
+    def atom_num(
+        self, mol: Chem.rdchem.Mol, cutoff: int = 4
+    ) -> Chem.rdchem.Mol or None:
+        """Filter molecules by number of atoms
+
+        Filter out molecules by the number of atoms.
+
+        Parameters
+        ----------
+        mol : Chem.rdchem.Mol
+          The Rdkit mol object of the input molecule.
+        cutoff : int
+          The criterion of atom numbers for filtering. By default 4.
+
+        Returns
+        -------
+        Chem.rdchem.Mol or None
+          Return a Rdkit mol object if the input molecule meets the criterion, otherwise return None.
+        """
+        atom_num = mol.GetNumAtoms()
[…]
