### 2023-09-04 b4adca0
Moved mol_filter.py to benchmark folder
