### 2023-09-04 #5763 @richardjgowers
@exs-cbouy yes if you could rerun the notebook in the repo and verify the results, i.e. a manual notebook regression test
