### 2023-09-04 efbd1a3
Refactored small modules into fepops.utils
