### 2023-09-04 5bd05d8
Removed pytest from build requirements
