### 2023-09-05 7951073
Replaced print error with logpping.error #18
