### 2023-09-05 0c83e6a
Replaced print warning msg with logpping.warning #18
