### 2023-09-05 7b929a1
Updated benchmarking notebook
