### 2023-09-05 72a794e
Updated JSON output code
