### 2023-09-05 #5763 @stevenshave
> @richardjgowers as a reviewer am I supposed to rerun the benchmarking notebook (as it would qualify as "original results") and make sure that the results are the same, or is making sure that the notebook is runnable enough? It's quite slow to run (but the notebook works).

Hi @exs-cbouy , thank you very much for your in depth review.  We are working on addressing all points.  As for rerunning the notebook, this has been refactored somewhat in addressing your comments. The most up to date version in the development branch (not yet merged to main) would be the best to use, available here: https://github.com/JustinYKC/FEPOPS/blob/7b929a16dc1eadfa1f4abfd4e80a4c2a7982cad9/Explore_DUDE_diversity_set.ipynb

In addition, pre-generated FEPOPS database files for the DUDE diversity set are available here: https://doi.org/10.6084/m9.figshare.23951445.v3 which will massively speedup running the notebook. These new descriptors have been generated after addressing your point on feature scaling before k-medoid clustering and so differ from earlier versions.

Once we have addressed all changes on the development branch we will merge to main and notify all here.

Many thanks,
Steve
