### 2023-09-12 c7620c0
Updated docs and readme with JSON output options
