+++ docsource/modules.rst
@@ -3,4 +3,3 @@
    :maxdepth: 4
 
    fepops
-   fepops.fepops_persistent
