+++ .github/workflows/documentation.yaml
@@ -1,3 +1,5 @@
+# Sphinx is pinned to 6.2.1, can be any really, but not 7.2.5 as this bug with Numpy
+# arose: https://github.com/numpy/numpy/issues/24604
 name: Docs
 on: [push, pull_request, workflow_dispatch]
 permissions:
@@ -10,7 +12,9 @@ jobs:
       - uses: actions/setup-python@v3
       - name: Install dependencies
         run: |
-          pip install sphinx sphinx_rtd_theme recommonmark myst_parser
+          python -m pip install --upgrade pip
+          pip install -e .[dev]
+          pip install sphinx==6.2.1 sphinx_rtd_theme recommonmark myst_parser
       - name: Sphinx build
         run: |
           sphinx-build docsource docs
