### 2023-09-13 JustinYKC/FEPOPS#16 @JustinYKC
We have expanded the paper (paper, readme and docs - 6280684620464c95da3f80ff766942eb1c86e472) with an explanation of how atomic properties are calculated with RDKit (ref added) – using the Gasteiger algorithm (reference added) for atomic partial charges, Crippen method (reference added) for atomic logP contributions and referenced the paper from which the HBA+HBD acceptor patterns were taken - https://pubs.acs.org/doi/10.1021/ci970431+ . This has been merged into main in 2146bcf0b2737ea671cb4beed6cbfc5632f8baee and is present in OpenFEPOPS v 1.8.0.
