### 2023-09-13 JustinYKC/FEPOPS#15 @JustinYKC
This has been corrected in fb7214bb07d07d4c049cc14141cfc14bc2ab4795 which was later merged to master in 2146bcf0b2737ea671cb4beed6cbfc5632f8baee then addressed again in f16d0e1c58005b330b7397fb8c9e1831165f7318 when a conflict arose in sphinx 7.2.5 with Numpy. Sphinx was then pinned to an earlier version. API docs now render correctly and support the OpenFEPOPS v1.8.0 release.
