### 2023-09-13 JustinYKC/FEPOPS#14 @JustinYKC
We have now implemented a CI workflow in f0a9f0952a2098d0d70345790c5d48000447cb25 for the master and development branches which runs the black formatter and tests upon push and pull requests. In addition, instructions for running tests have been added to the readme and online documentation. This has been merged into master in 2146bcf0b2737ea671cb4beed6cbfc5632f8baee and is present in OpenFEPOPS v 1.8.0.
