### 2023-09-13 JustinYKC/FEPOPS#11 @JustinYKC
Many thanks for spotting this bug. It has now been addressed in 0540c88f204bc45d9ae526225c96b90289493f6c and scaled features are used to assign representative medoids. This did have the effect of changing selected medoids and required regeneration of DUDE diversity set datasets, recapture of means + stds (changes in the order of 1e-3 per mean and 6e-4 for stds) for use in scoring, and a subsequent rerun of benchmarking (now updated in the paper, readme and docs) in 199ee51e2fe595f3ea975db27e32d6c4230947b5. This has been merged into main in 2146bcf0b2737ea671cb4beed6cbfc5632f8baee and is present in OpenFEPOPS v 1.8.0.
