### 2023-09-14 #5763 @stevenshave
Hi @richardjgowers and @exs-cbouy, we have now addressed the raised issues and released a new version of OpenFEPOPS (v.1.8.0) which includes bugfixes and updates to functionality, online documentation and the manuscript.
Many thanks,
Steve
