### 2023-09-15 JustinYKC/FEPOPS#16 @cbouysset
There seem to be an issue with the reference for Gasteiger charges, it should be [Gasteiger & Marsili](https://doi.org/10.1016/0040-4020(80)80168-2) instead of Wildman & Crippen.
