### 2023-09-15 JustinYKC/FEPOPS#14 @cbouysset
Just a quick note (optional to address) on the CI workflow, on [these lines](https://github.com/JustinYKC/FEPOPS/blob/master/.github/workflows/ci.yaml#L6-L8) the target branch should most likely be `master` instead of `main` (unless you plan on renaming the master branch to main).
