### 2023-09-15 debfe34
Corrected CI workflow to operate on master branch
