+++ paper.bib
@@ -154,3 +154,15 @@
   publisher={ACS Publications},
   doi={10.1021/ci990307l}
 }
+
+@article{gasteiger1980iterative,
+  title={Iterative partial equalization of orbital electronegativity—a rapid access to atomic charges},
+  author={Gasteiger, Johann and Marsili, Mario},
+  journal={Tetrahedron},
+  volume={36},
+  number={22},
+  pages={3219--3228},
+  year={1980},
+  publisher={Elsevier},
+  doi={10.1016/0040-4020(80)80168-2}
+}
