### 2023-09-15 JustinYKC/FEPOPS#16 @stevenshave
Thank you, this was certainly an error on my part. The Gasteiger reference has now been properly inserted in 8303a1815f5a176d6c9c6aefcf4d01035d4ee333.
