### 2023-09-18 JustinYKC/FEPOPS#20 @exs-cbouy
[Reproducibility] 6 of the 8 targets are skipped in benchmark notebook
Refers to the JOSS submission review: https://github.com/openjournals/joss-reviews/issues/5763

In the notebook, the first cell of the `Benchmark FEPOPS using the DUDE diversity set` section (third cell of the nb), the following change should be made:

```diff
diversity_set_csv_files = list(Path("data/dude/processed/diversity_set/")
-    .glob("dude_target_hiv*.csv")
+    .glob("dude_target*.csv")
)
```

Otherwise the notebook will only run for 2 out of 8 targets. Also the output of the same cell prints:
```
Got 1 diversity_set_csv_files : ['dude_target_ampc']
```
Although it should print out the 8 targets to be complete.

After making the above change and running the notebook I managed to reproduce the results shown in the table.
