### 2023-09-20 a13a7de
Addressed reproducibility issues in notebook #20
