### 2023-09-20 JustinYKC/FEPOPS#20 @stevenshave
Thank you for being so thorough with the review, and apologies that elements from testing made it into the notebook. This has now been addressed in a13a7de67e2b7f1da8d38eba616432744c15fbf3 and committed to the master branch with the fixed file glob pattern, and a refresh of all cell outputs.
