### 2023-09-20 873f200
Added kmeans_method arg to storgeobject getter #18
