### 2023-09-21 #5763 @exs-cbouy
Thank you @stevenshave for this open-source version of FEPOPS and the benchmark with commonly used fingerprints, I believe this will be a useful 3D representation for a variety of use cases, as it becomes more widely available to the public.

@richardjgowers All my comments have been addressed and my checklist is complete, LGTM!
