### 2023-10-09 JustinYKC/FEPOPS#22 @hannahbaumann
[ Software paper ] minor suggestions
Refers to the JOSS submission review: https://github.com/openjournals/joss-reviews/issues/5763:

Some small suggestions on the software paper:

- For the acronyms FEPOPS and AUROC it would be helpful to add a brief definition/spell out the acronyms upon first mentioning in the text
- line 84: typo: realatively --> relatively
- line 110: Add a brief explanation on how the macro-averaged AUROC scores were computed.
- results part (lines 108 and following): Add a sentence briefly comparing the performance of the different methods (Table 1)
