### 2023-10-10 3eeca2f
Corrected rdkit dependency
Mixture of rdkit-pypi and rdkit standardized to just "rdkit".
