+++ requirements.txt
@@ -2,7 +2,7 @@ fast-pytorch-kmeans==0.1.9
 fire==0.5.0
 numpy==1.22.4
 pandas==1.5.0
-rdkit-pypi==2021.9.5.1
+rdkit==2022.3.1
 scikit-learn==1.3.0
 scipy==1.7.0
 torch==1.7.1+cu101
