+++ paper.bib
@@ -166,3 +166,15 @@
   publisher={Elsevier},
   doi={10.1016/0040-4020(80)80168-2}
 }
+
+@article{fawcett2006introduction,
+  title={An introduction to ROC analysis},
+  author={Fawcett, Tom},
+  journal={Pattern recognition letters},
+  volume={27},
+  number={8},
+  pages={861--874},
+  year={2006},
+  publisher={Elsevier},
+  doi={10.1016/j.patrec.2005.10.010}
+}
