### 2023-10-11 7ec66ca
Removed macro from averaging scheme
