### 2023-10-11 JustinYKC/FEPOPS#22 @stevenshave
Thank you for your comments and suggestions, we have now integrated all of them into the submission.

Typo addressed and acronyms are listed in full upon their first occurrence and we have added a citation breaking down the AUROC and ROC metrics.

We have removed the term macro from "macro averaging" throughout the paper, the README.md ,and online docs as it was not necessary here. We have clarified the process used in averaging AUROC scores in the paper text.

In addition, we have a short comparative summary of metric performance and conclusion.

Many thanks,
Steve
