### 2023-10-11 #5763 @stevenshave
Thank you for your review @hannahbaumann , we believe we have addressed all comments and fixed the installation issue in the new 1.8.1 OpenFEPOPS release which is now on pypi.

Best,
Steve
