### 2023-10-13 JustinYKC/FEPOPS#23 @hannahbaumann
Error when running "Explore_DUDE_diversity_set.ipynb"
Refers to the JOSS submission review: https://github.com/openjournals/joss-reviews/issues/5763:

I was able to run most of the example notebook, but for some reason my `res_df_dude_target_*.csv` files look different compared to the original files that were provided in the example, leading to an error when trying to generate the tables from the paper using this command:
```
display(pd.concat([pd.read_csv(f) for f in Path("data/dude/processed/diversity_set/").glob("res_df_dude_target_hivrt.csv")])
        .pivot(values='average_auroc_score',index='target', columns=['similarity_method'],)
        [['Morgan 2', 'RDKit', 'MACCS', 'OpenFEPOPS']])
```

The error is the following:

```
File ~/mambaforge/envs/fepops_311/lib/python3.10/site-packages/pandas/core/reshape/reshape.py:210, in _Unstacker._make_selectors(self)
    207 mask.put(selector, True)
    209 if mask.sum() < len(self.index):
--> 210     raise ValueError("Index contains duplicate entries, cannot reshape")
    212 self.group_index = comp_index
    213 self.mask = mask

ValueError: Index contains duplicate entries, cannot reshape
```

Even though the code that generates the .csv files for the different targets ran successfully without giving an error (calculating the different metrics for the 4 different methods), the output files are missing the Morgan 2 method, and instead have the row for the MACCS method duplicated which then leads to the error when trying to generate the table.

Output example: `res_df_dude_target_cp3a4.csv`

```
,target,similarity_method,smiles_count,actives_count,failed_smiles,failed_active_smiles,average_auroc_score,median_auroc_score,q1_auroc_score,q3_auroc_score
0,cp3a4,MACCS,11970,170,4,0,0.5815453804644352,0.5808045598707439,0.42871445517423656,0.45669245058145336
1,cp3a4,MACCS,11970,170,4,0,0.5815453804644352,0.5808045598707439,0.42871445517423656,0.45669245058145336
2,cp3a4,RDKit,11970,170,4,0,0.6133354937630608,0.6263335776833623,0.3858205965880757,0.3903332722458261
3,cp3a4,OpenFEPOPS,11970,170,4,0,0.6474616432055439,0.653921568627451,0.36536023801687506,0.391133003710131
```

I had run the example notebook as provided on the github page, but had skipped the first step of preparing the DUDE database and instead used the files provided in `data/dude/processed/diversity_set`
Do you have an idea what could be going on here? Please let me know if it would be helpful if I provided more details!
