### 2023-10-16 JustinYKC/FEPOPS#23 @stevenshave
Hi @hannahbaumann, thank you for raising this issue.

I found an issue on my side with the jupyter notebook which contained a link pointing to version 2 of the data (which was incorrect as you note). Version 3 (released  2023-09-05) should be used.  Apologies for that, I have now updated the link in the notebook to point to v3 (https://doi.org/10.6084/m9.figshare.23951445.v3) in commit ef17f328ce58498828a11c8147b307d45308aa1e.

Another possible issue - can you confirm that you are using the latest version of the jupyter notebook (and/or redownload the new one with the corrected link to FigShare v3 files)? In your text above, you note that the cell which throws the error has glob string of `res_df_dude_target_hivrt.csv`, but in the current version it has the string `res_df_dude_target_*.csv`.  If you had changed it just for your testing purposes, then possibly the pivot is failing due to only one DataFrame being loaded.

The notebook works for me using V3 data and the most up-to-date notebook. I have :
1) Re-downloaded the V3 data zip file and rerun the notebook from the title `# Output summary AUROC results table as shown in the paper`. This reproduced the table in the paper.
2) Re-run the entire notebook including descriptor generation. This reproduced the table in the paper.

Please let me know if this fixes the issues.
Many thanks,
Steve
