### 2023-10-18 JustinYKC/FEPOPS#23 @hannahbaumann
Hi @stevenshave : Thank you so much for your help!
I re-downloaded the data zip file (making sure to download v3) and I also re-downloaded the jupyter notebook from the master branch.
I was able to generate the tables from the paper when using the files form the data folder and only running the notebook from the title `# Output summary AUROC results table as shown in the paper`
However, when rerunning the entire notebook including descriptor generation I was still getting the same error. Something weird happens when generating the `auroc_scores_info_df`. In the first round of the loop
```
info=pd.Series(dtype=object)
for sm_name, sm in similarity_methods.items():
        # Remove entries which did not return a mol
        info['target']=csv_file_path.stem.replace("dude_target_","")
```
the code correctly generates the dataframe for Morgan2, but then when calculating values for MACCS it replaces the Morgan2 column with MACCS and ends up having two duplicate MACCS columns, one rdkit and one openfepops column. I'll keep looking into this, but if you already have an idea of what could be going on on my side, please let me know!
