### 2023-10-31 3d28dd8
Benchmarking aborts when results files exist #23
