### 2023-10-31 JustinYKC/FEPOPS#23 @stevenshave
Hi @hannahbaumann, I've rewritten the benchmark code somewhat and included some guards for results files existing in the current directory, along with changes to the loop which populates results files.  This works with me running it from scratch, and also with the intermediate files in the data zip (V3-as before).

Would you be able to re-download the notebook and check that the new version works for you.
Many thanks,
Steve
