### 2023-10-31 c1e0564
Updated rdkit requirements
