+++ pyproject.toml
@@ -10,7 +10,7 @@ readme = "README.md"
 requires-python = ">=3.9"
 keywords = ["molecular similarity", "ligand based virtual screening"]
 dependencies = [
-    "rdkit>=2022.3.1",
+    "rdkit>=2022.3.1, <2023.9.1",
     "numpy>=1.22.4",
     "pandas>=1.5.0",
     "scikit-learn>=1.3.0",
