### 2023-11-01 786424c
Bump version to 1.8.2
