+++ .github/workflows/documentation.yaml
@@ -10,6 +10,8 @@ jobs:
     steps:
       - uses: actions/checkout@v3
       - uses: actions/setup-python@v3
+        with:
+          python-version: '3.11' 
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
