### 2023-11-01 #5763 @richardjgowers
Hi @stevenshave we're onto the final checklist above.  Can you make a zenodo/figshare release of 1.8.2 and post the DOI here?
