### 2023-11-02 #5763 @stevenshave
Hi @richardjgowers, I've just added the 1.8.2 source archive to figshare under the URL:
https://doi.org/10.6084/m9.figshare.24477184.v1

and DOI:
10.6084/m9.figshare.24477184.v1

Many thanks to you and the reviewers for progressing this!
Steve
