### 2023-11-08 #5763 @Kevin-Mattheus-Moerman
@stevenshave as the AEiC on this track I will now help to process the final steps. I have checked the repository, this review, the paper, and the archive link. Most seems in order. However I have the below points that require your attention:

On the paper:
- [ ] Both British English (e.g. `neighbouring`) and American English (e.g. `featurization`) are used. Please correct to be consistent.
- [ ] Check spelling for `totalling`, which should perhaps be `totaling`.

Comments/recommendations (not required):
- [ ] Your affiliations are currently (nearly) full postal addresses. However, if you prefer to shorten this you may do so, e.g. to simply have the format <Department>, <University/Institute/Company>, <City>, <Country>. In other words, you may, if you like, leave out street/building names and postal/zip codes.
- [ ] For your first reference (K-means++ the advantages of careful seeding) you may consider adding a `url` entry in your bib file e.g. to: https://dl.acm.org/doi/abs/10.5555/1283383.1283494 (looks like their DOI does not resolve but this url does).
