### 2023-11-08 6657d13
Spelling alterations
"Totalling" to "Totaling"
"Featurize" to "Featurise"
also found "Fingeprint", changed to "Fingerprint"
