### 2023-11-08 6648132
Added URL to arthur2007k reference
