+++ paper.bib
@@ -116,7 +116,8 @@
   author={Arthur, David and Vassilvitskii, Sergei},
   booktitle={Proceedings of the eighteenth annual ACM-SIAM symposium on Discrete algorithms},
   pages={1027--1035},
-  year={2007}
+  year={2007},
+  url={https://dl.acm.org/doi/abs/10.5555/1283383.1283494}
 }
 
 @article{park2009simple,
