### 2023-11-08 #5763 @stevenshave
Hi @Kevin-Mattheus-Moerman, many thanks for your work on this.

I have now standardised to British English, replacing featurization with featurisation. I looked into the totalling vs totaling and it seems unclear, with both being acceptable, but have gone with your suggestion of totaling.  I discovered another typo of "fingeprint" which has now been corrected.

After a brief discussion with coauthors, one would strongly prefer keeping the addresses as prescribed by our departments for use in publishing.  I hope that it is OK if these remain as they are.

I have added the suggested URL to the Arthur 2007 reference, which now appears correctly.

My thanks to all reviewers and editors for the work you have all contributed to this, it has been a great experience!

Best,
Steve
