### 2023-11-09 146d479
Added CITATION.cff after acceptance to JOSS
