+++ _sources/fepops.fepops_persistent.rst.txt
@@ -1,0 +1,45 @@
+fepops.fepops\_persistent package
+=================================
+
+Submodules
+----------
+
+fepops.fepops\_persistent.fepops\_persistent\_abc module
+--------------------------------------------------------
+
+.. automodule:: fepops.fepops_persistent.fepops_persistent_abc
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+fepops.fepops\_persistent.fepopsdb\_json module
+-----------------------------------------------
+
+.. automodule:: fepops.fepops_persistent.fepopsdb_json
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+fepops.fepops\_persistent.fepopsdb\_sqlite module
+-------------------------------------------------
+
+.. automodule:: fepops.fepops_persistent.fepopsdb_sqlite
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+fepops.fepops\_persistent.utils module
+--------------------------------------
+
+.. automodule:: fepops.fepops_persistent.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: fepops.fepops_persistent
+   :members:
+   :undoc-members:
+   :show-inheritance:
