### 2023-11-09 #5763 @stevenshave
I see that the paper is now live!

I have added a CITATION.cff to the repository and GitHub has picked it up, enabling the "Cite this repository" link/button.

Again, my thanks to the reviewers for the in depth reviews which forced me to rethink many things and also become more proficient at github, and to the editors for making this such a good experience.  We will be looking in the future to improve the speed of the technique with some compiled rust to replace performance critical sections and increase the utility of OpenFEPOPS.

Best wishes,
Steve
